add Phase 1: HDr and Metalog Wrappers

+++ modules/common/src/main/scala/com/risquanter/register/http/responses/SimulationResponse.scala
@@ -27,6 +27,6 @@ object SimulationResponse {
     minLoss = sim.minLoss,
     maxLoss = sim.maxLoss,
     likelihoodId = sim.likelihoodId,
-    probability = sim.probability
+    probability = sim.probability.toDouble
   )
 }
